+++ cql3/attributes.cc
@@ -10,6 +10,8 @@
 
 #include "cql3/attributes.hh"
 #include "cql3/column_identifier.hh"
+#include "cql3/expr/evaluate.hh"
+#include "cql3/expr/expr-utils.hh"
 #include <optional>
 
 namespace cql3 {
