+++ cql3/statements/cas_request.cc
@@ -13,6 +13,8 @@
 #include "cas_request.hh"
 #include <seastar/core/sleep.hh>
 #include "cql3/result_set.hh"
+#include "cql3/expr/evaluate.hh"
+#include "cql3/expr/expr-utils.hh"
 #include "transport/messages/result_message.hh"
 #include "types/map.hh"
 #include "service/storage_proxy.hh"
