+++ cql3/statements/strongly_consistent_modification_statement.cc
@@ -20,6 +20,7 @@
 #include "bytes.hh"
 #include "cql3/attributes.hh"
 #include "cql3/expr/expression.hh"
+#include "cql3/expr/evaluate.hh"
 #include "cql3/operation.hh"
 #include "cql3/query_processor.hh"
 #include "cql3/values.hh"
