+++ cql3/statements/strongly_consistent_select_statement.cc
@@ -15,6 +15,7 @@
 #include <seastar/core/on_internal_error.hh>
 
 #include "cql3/restrictions/statement_restrictions.hh"
+#include "cql3/expr/evaluate.hh"
 #include "cql3/query_processor.hh"
 #include "service/broadcast_tables/experimental/lang.hh"
 #include "db/system_keyspace.hh"
