+++ cql3/update_parameters.cc
@@ -11,6 +11,8 @@
 #include "cql3/update_parameters.hh"
 #include "cql3/selection/selection.hh"
 #include "cql3/expr/expression.hh"
+#include "cql3/expr/evaluate.hh"
+#include "cql3/expr/expr-utils.hh"
 #include "query-result-reader.hh"
 #include "types/map.hh"
