+++ test/boost/expr_test.cc
@@ -16,6 +16,8 @@
 #include "types/set.hh"
 #include "types/user.hh"
 #include "test/lib/expr_test_utils.hh"
+#include "cql3/expr/evaluate.hh"
+#include "cql3/expr/expr-utils.hh"
 
 using namespace cql3;
 using namespace cql3::expr;
