+++ test/boost/statement_restrictions_test.cc
@@ -12,6 +12,7 @@
 #include <vector>
 
 #include "cql3/restrictions/statement_restrictions.hh"
+#include "cql3/expr/expr-utils.hh"
 #include "cql3/util.hh"
 #include "test/lib/cql_assertions.hh"
 #include "test/lib/cql_test_env.hh"
